refactoring to create gallery component
